requirements.txt
altair==5.5.0
attrs==25.3.0
beautifulsoup4==4.13.3
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
esda==2.7.0
fonttools==4.56.0
geographiclib==2.0
geopandas==1.0.1
geopy==2.4.1
gitdb==4.0.12
GitPython==3.1.44
idna==3.10
imageio==2.37.0
Jinja2==3.1.6
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
lazy_loader==0.4
libpysal==4.12.1
lightgbm==4.6.0
lime==0.2.0.1
MarkupSafe==3.0.2
matplotlib==3.10.1
narwhals==1.31.0
networkx==3.4.2
numpy==2.2.4
packaging==24.2
pandas==2.2.3
patsy==1.0.1
pillow==11.1.0
platformdirs==4.3.7
plotly==6.0.1
protobuf==5.29.4
pyarrow==19.0.1
pydeck==0.9.1
pyogrio==0.10.0
pyparsing==3.2.2
pyproj==3.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.1
referencing==0.36.2
requests==2.32.3
rpds-py==0.23.1
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
shapely==2.0.7
six==1.17.0
smmap==5.0.2
soupsieve==2.6
statsmodels==0.14.4
streamlit==1.43.2
streamlit-option-menu==0.4.0
tenacity==9.0.0
threadpoolctl==3.6.0
tifffile==2025.3.13
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.3.0
watchdog==6.0.0
xgboost==3.0.0
